+++ Classes/BuildingObject.h
@@ -19,8 +19,8 @@ namespace rtm {
     {
     public:
         BuildingObject();
-        BuildingObject(std::string const& filename, size_t row, size_t column, float angle);
-        BuildingObject(BuildingType type, size_t row, size_t column, float angle);
+        BuildingObject(std::string const& filename, int row, int column, float angle);
+        BuildingObject(BuildingType type, int row, int column, float angle);
 
         virtual ~BuildingObject() = default;
